Create a file for Message/OutputLine
