+++ pylint/testutils/utils.py
@@ -20,7 +20,7 @@ from typing import Tuple
 import astroid
 import pytest
 
-from pylint import checkers, interfaces
+from pylint import checkers
 from pylint.lint import PyLinter
 from pylint.testutils.constants import (
     _EXPECTED_RE,
@@ -33,6 +33,7 @@ from pylint.testutils.functional_test_file import (
     NoFileError,
     parse_python_version,
 )
+from pylint.testutils.output_line import Message, OutputLine
 from pylint.testutils.reporter_for_tests import (
     FunctionalTestReporter,
     GenericTestReporter,
@@ -76,22 +77,6 @@ def _get_tests_info(input_dir, msg_dir, prefix, suffix):
             outfile = join(msg_dir, fbase + ".txt")
         result.append((infile, outfile))
     return result
-
-
-class Message(
-    collections.namedtuple("Message", ["msg_id", "line", "node", "args", "confidence"])
-):
-    def __new__(cls, msg_id, line=None, node=None, args=None, confidence=None):
-        return tuple.__new__(cls, (msg_id, line, node, args, confidence))
-
-    def __eq__(self, other):
-        if isinstance(other, Message):
-            if self.confidence and other.confidence:
-                return super().__eq__(other)
-            return self[:-1] == other[:-1]
-        return NotImplemented  # pragma: no cover
-
-    __hash__ = None
 
 
 class UnittestLinter:
@@ -229,35 +214,6 @@ def _create_file_backed_module(code):
         module = astroid.parse(code)
         module.file = temp
         yield module
-
-
-class OutputLine(
-    collections.namedtuple(
-        "OutputLine", ["symbol", "lineno", "object", "msg", "confidence"]
-    )
-):
-    @classmethod
-    def from_msg(cls, msg):
-        return cls(
-            msg.symbol,
-            msg.line,
-            msg.obj or "",
-            msg.msg.replace("\r\n", "\n"),
-            msg.confidence.name
-            if msg.confidence != interfaces.UNDEFINED
-            else interfaces.HIGH.name,
-        )
-
-    @classmethod
-    def from_csv(cls, row):
-        confidence = row[4] if len(row) == 5 else interfaces.HIGH.name
-        return cls(row[0], int(row[1]), row[2], row[3], confidence)
-
-    def to_csv(self):
-        if self.confidence == interfaces.HIGH.name:
-            return self[:-1]
-
-        return self
 
 
 def parse_expected_output(stream):
